requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
dnspython>=2.1.0
email-validator>=1.1.3 